separate unused weeklyscheduleview
